tools/blog-editor/pyproject.toml
[project]
name = "blog-editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0"
]
